requirements.txt
clearvoice
pyannote.audio>=3.3
aiohttp
bs4